:truck: move CertificationComputeError to certif contexte
