+++ api/src/certification/evaluation/domain/usecases/score-completed-v2-certification.js
@@ -7,8 +7,8 @@
  * @typedef {import('./index.js').ComplementaryCertificationScoringCriteriaRepository} ComplementaryCertificationScoringCriteriaRepository
  */
 
-import { CertificationComputeError } from '../../../../shared/domain/errors.js';
 import { AssessmentResultFactory } from '../../../scoring/domain/models/factories/AssessmentResultFactory.js';
+import { CertificationComputeError } from '../errors.js';
 
 /**
  * @param {Object} params
